Refactor the AddCommand to use cloud
